Added stereo calibration using charuco board (#976)
From #972
Doing this first for rolling.

This was a TODO in the repository, opening this PR to add this feature.
- The main issue why this wasn't possible imo is the way `mk_obj_points`
works. I'm using the inbuilt opencv function to get the points there.
- The other is a condition when aruco markers are detected they are
added as good points, This is fine in case of mono but in stereo these
have to be the same number as the object points to find matches although
this should be possible with aruco.

(cherry picked from commit efb9005a5ca45f4ceb7563fc01d7e33d6f39214d)

# Conflicts:
#	camera_calibration/src/camera_calibration/calibrator.py